requirements.txt
pandas
numpy
matplotlib
seaborn
paho-mqtt
flask
requests
scikit-learn